Update integration tests to use latest Vault

+++ test/bats/provider.bats
@@ -13,13 +13,15 @@ CONFIGS=test/bats/configs
 setup(){
     { # Braces used to redirect all setup logs.
     # 1. Configure Vault.
-    kubectl --namespace=csi exec vault-0 -- vault namespace create acceptance
 
     # 1. a) Vault policies
     cat $CONFIGS/vault-policy-db.hcl | kubectl --namespace=csi exec -i vault-0 -- vault policy write db-policy -
     cat $CONFIGS/vault-policy-kv.hcl | kubectl --namespace=csi exec -i vault-0 -- vault policy write kv-policy -
     cat $CONFIGS/vault-policy-pki.hcl | kubectl --namespace=csi exec -i vault-0 -- vault policy write pki-policy -
-    cat $CONFIGS/vault-policy-kv-namespace.hcl | kubectl --namespace=csi exec -i vault-0 -- vault policy write -namespace=acceptance kv-namespace-policy -
+    if [ -n "${VAULT_LICENSE}" ]; then
+        kubectl --namespace=csi exec vault-0 -- vault namespace create acceptance
+        cat $CONFIGS/vault-policy-kv-namespace.hcl | kubectl --namespace=csi exec -i vault-0 -- vault policy write -namespace=acceptance kv-namespace-policy -
+    fi
     cat $CONFIGS/vault-policy-kv-custom-audience.hcl | kubectl --namespace=csi exec -i vault-0 -- vault policy write kv-custom-audience-policy -
 
     # 1. b) Setup kubernetes auth engine.
@@ -30,16 +32,12 @@ setup(){
     # For AKS: "<dns-prefix>.hcp.<region>.azmk8s.io" (same as API server address, but with quotes)
     # For GKE: https://container.googleapis.com/v1/projects/<project>/locations/<az|region>/clusters/<cluster-name>
     kubectl --namespace=csi exec vault-0 -- sh -c 'vault write auth/kubernetes/config \
-        token_reviewer_jwt="$(cat /var/run/secrets/kubernetes.io/serviceaccount/token)" \
-        kubernetes_host="https://$KUBERNETES_PORT_443_TCP_ADDR:443" \
-        kubernetes_ca_cert=@/var/run/secrets/kubernetes.io/serviceaccount/ca.crt \
-        issuer="https://kubernetes.default.svc.cluster.local"'
-    kubectl --namespace=csi exec vault-0 -- vault auth enable -namespace=acceptance kubernetes
-    kubectl --namespace=csi exec vault-0 -- sh -c 'vault write -namespace=acceptance auth/kubernetes/config \
-        token_reviewer_jwt="$(cat /var/run/secrets/kubernetes.io/serviceaccount/token)" \
-        kubernetes_host="https://$KUBERNETES_PORT_443_TCP_ADDR:443" \
-        kubernetes_ca_cert=@/var/run/secrets/kubernetes.io/serviceaccount/ca.crt \
-        issuer="https://kubernetes.default.svc.cluster.local"'
+        kubernetes_host="https://$KUBERNETES_PORT_443_TCP_ADDR:443"'
+    if [ -n "${VAULT_LICENSE}" ]; then
+        kubectl --namespace=csi exec vault-0 -- vault auth enable -namespace=acceptance kubernetes
+        kubectl --namespace=csi exec vault-0 -- sh -c 'vault write -namespace=acceptance auth/kubernetes/config \
+            kubernetes_host="https://$KUBERNETES_PORT_443_TCP_ADDR:443"'
+    fi
     kubectl --namespace=csi exec vault-0 -- vault write auth/kubernetes/role/db-role \
         bound_service_account_names=nginx-db \
         bound_service_account_namespaces=test \
@@ -58,12 +56,14 @@ setup(){
         bound_service_account_namespaces=test \
         policies=kv-custom-audience-policy \
         ttl=20m
-    kubectl --namespace=csi exec vault-0 -- vault write -namespace=acceptance auth/kubernetes/role/kv-namespace-role \
-        bound_service_account_names=nginx-kv-namespace \
-        bound_service_account_namespaces=test \
-        audience=vault \
-        policies=kv-namespace-policy \
-        ttl=20m
+    if [ -n "${VAULT_LICENSE}" ]; then
+        kubectl --namespace=csi exec vault-0 -- vault write -namespace=acceptance auth/kubernetes/role/kv-namespace-role \
+            bound_service_account_names=nginx-kv-namespace \
+            bound_service_account_namespaces=test \
+            audience=vault \
+            policies=kv-namespace-policy \
+            ttl=20m
+    fi
     kubectl --namespace=csi exec vault-0 -- vault write auth/kubernetes/role/pki-role \
         bound_service_account_names=nginx-pki \
         bound_service_account_namespaces=test \
@@ -94,8 +94,10 @@ setup(){
     kubectl --namespace=csi exec vault-0 -- vault kv put secret/kv-sync1 bar1=hello-sync1
     kubectl --namespace=csi exec vault-0 -- vault kv put secret/kv-sync2 bar2=hello-sync2
     kubectl --namespace=csi exec vault-0 -- vault kv put secret/kv-sync3 bar3=aGVsbG8tc3luYzM=
-    kubectl --namespace=csi exec vault-0 -- vault secrets enable -namespace=acceptance -path=secret -version=2 kv
-    kubectl --namespace=csi exec vault-0 -- vault kv put -namespace=acceptance secret/kv1-namespace greeting=hello-namespaces
+    if [ -n "${VAULT_LICENSE}" ]; then
+        kubectl --namespace=csi exec vault-0 -- vault secrets enable -namespace=acceptance -path=secret -version=2 kv
+        kubectl --namespace=csi exec vault-0 -- vault kv put -namespace=acceptance secret/kv1-namespace greeting=hello-namespaces
+    fi
     kubectl --namespace=csi exec vault-0 -- vault kv put secret/kv-custom-audience bar=hello-custom-audience
 
     # 2. Create shared k8s resources.
@@ -130,7 +132,9 @@ teardown(){
     fi
 
     # Teardown Vault configuration.
-    kubectl --namespace=csi exec vault-0 -- vault namespace delete acceptance
+    if [ -n "${VAULT_LICENSE}" ]; then
+        kubectl --namespace=csi exec vault-0 -- vault namespace delete acceptance
+    fi
     kubectl --namespace=csi exec vault-0 -- vault auth disable kubernetes
     kubectl --namespace=csi exec vault-0 -- vault secrets disable secret
     kubectl --namespace=csi exec vault-0 -- vault secrets disable pki
@@ -325,6 +329,9 @@ teardown(){
 }
 
 @test "9 Vault Enterprise namespace" {
+    if [ -z "${VAULT_LICENSE}" ]; then
+        skip "No Vault license configured, skipping namespace test"
+    fi
     helm --namespace=test install nginx $CONFIGS/nginx \
         --set engine=kv-namespace --set sa=kv-namespace \
         --wait --timeout=5m
